refactor: move watcher and nlp bot into one directory
the projects are now treated as sub modules of the telegram_bots module. This will make it easier to share code between the projects and keep the telegram bot more maintainable.
