+++ packages/telegram_bots/FreiFahren_BE_NLP/data/stations_and_lines.json
