+++ packages/telegram_bots/FreiFahren_BE_NLP/data/synonyms.json
