--- packages/watcher/setup.cfg
@@ -1,20 +1,0 @@
-[flake8]
-# General
-max-line-length = 180
-exclude = .git,__pycache__,docs/source/conf.py,old,build,dist,venv,.json
-ignore = E203, E266, W503, W293, W504
-max-complexity = 9
-show-source = True
-statistics = True
-
-# Ignore specific errors in specified files
-per-file-ignores =
-    test_cases.py:E121,E501 remove_direction_and_keyword_test.py:N802 test_main.py:N802, E402  ner_test_cases.py: E121, E501 ner_performance_test.py: N802, E402 general_accuracy_test.py: E402
-
-# Compatibility with Black
-extend-ignore = E203, W503
-
-# Plugin: flake8-quotes
-inline-quotes = '
-multiline-quotes = '''
-docstring-quotes = '''
